Consolidate Tx and RWTx.
