+++ example_test.go
@@ -67,7 +67,7 @@ func ExampleDB_Do() {
 	defer db.Close()
 
 	// Execute several commands within a write transaction.
-	err := db.Do(func(t *RWTx) error {
+	err := db.Do(func(t *Tx) error {
 		if err := t.CreateBucket("widgets"); err != nil {
 			return err
 		}
@@ -134,30 +134,30 @@ func ExampleDB_ForEach() {
 	// A liger is awesome.
 }
 
-func ExampleRWTx() {
+func ExampleTx() {
 	// Open the database.
 	var db DB
-	db.Open("/tmp/bolt/rwtx.db", 0666)
+	db.Open("/tmp/bolt/tx.db", 0666)
 	defer db.Close()
 
 	// Create a bucket.
 	db.CreateBucket("widgets")
 
 	// Create several keys in a transaction.
-	rwtxn, _ := db.RWTx()
-	b := rwtxn.Bucket("widgets")
+	tx, _ := db.RWTx()
+	b := tx.Bucket("widgets")
 	b.Put([]byte("john"), []byte("blue"))
 	b.Put([]byte("abby"), []byte("red"))
 	b.Put([]byte("zephyr"), []byte("purple"))
-	rwtxn.Commit()
+	tx.Commit()
 
 	// Iterate over the values in sorted key order.
-	txn, _ := db.Tx()
-	c := txn.Bucket("widgets").Cursor()
+	tx, _ = db.Tx()
+	c := tx.Bucket("widgets").Cursor()
 	for k, v := c.First(); k != nil; k, v = c.Next() {
 		fmt.Printf("%s likes %s\n", string(k), string(v))
 	}
-	txn.Close()
+	tx.Rollback()
 
 	// Output:
 	// abby likes red
@@ -165,10 +165,10 @@ func ExampleRWTx() {
 	// zephyr likes purple
 }
 
-func ExampleRWTx_rollback() {
+func ExampleTx_rollback() {
 	// Open the database.
 	var db DB
-	db.Open("/tmp/bolt/rwtx_rollback.db", 0666)
+	db.Open("/tmp/bolt/tx_rollback.db", 0666)
 	defer db.Close()
 
 	// Create a bucket.
@@ -178,10 +178,10 @@ func ExampleRWTx_rollback() {
 	db.Put("widgets", []byte("foo"), []byte("bar"))
 
 	// Update the key but rollback the transaction so it never saves.
-	rwtxn, _ := db.RWTx()
-	b := rwtxn.Bucket("widgets")
+	tx, _ := db.RWTx()
+	b := tx.Bucket("widgets")
 	b.Put([]byte("foo"), []byte("baz"))
-	rwtxn.Rollback()
+	tx.Rollback()
 
 	// Ensure that our original value is still set.
 	value, _ := db.Get("widgets", []byte("foo"))
